requirements.txt
Django==2.2.10
django-ckeditor==6.0.0
django-debug-toolbar==2.2
django-js-asset==1.2.2
django-solo==1.1.3
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.0
